GCP_projects/GCP_FlaskMarket/requirements.txt
aiohttp
aiosignal
APScheduler
asn1crypto
async-timeout
attrs
Beaker
blinker
cachelib
cachetools
certifi
cffi
charset-normalizer
click
cloud-sql-python-connector
colorama
cryptography
Flask
Flask-APScheduler
Flask-SQLAlchemy
Flask-WTF
frozenlist
google-api-core
google-auth
google-cloud-secret-manager
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
multidict
packaging
pg8000
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
python-dateutil
pytz
requests
rsa
scramp
six
SQLAlchemy
typing_extensions
tzdata
tzlocal
urllib3
Werkzeug
WTForms
yarl
